+++ packages/commonwealth/client/scripts/views/pages/CreateCommunity/CreateCommunity.tsx
@@ -1,64 +1,38 @@
-import React, { useState } from 'react';
+import React from 'react';
 
-import AddressInfo from 'models/AddressInfo';
-import { SelectedCommunity } from 'views/components/component_kit/new_designs/CWCommunitySelector';
 import CWFormSteps from 'views/components/component_kit/new_designs/CWFormSteps';
 
 import { MixpanelCommunityCreationEvent } from '../../../../../shared/analytics/types';
 import { useBrowserAnalyticsTrack } from '../../../hooks/useBrowserAnalyticsTrack';
 import BasicInformationStep from './steps/BasicInformationStep';
-import { ETHEREUM_MAINNET_ID } from './steps/BasicInformationStep/BasicInformationForm/constants';
 import CommunityStakeStep from './steps/CommunityStakeStep';
 import CommunityTypeStep from './steps/CommunityTypeStep';
 import SuccessStep from './steps/SuccessStep';
-import { CreateCommunityStep, getFormSteps, handleChangeStep } from './utils';
+import useCreateCommunity from './useCreateCommunity';
+import { CreateCommunityStep, getFormSteps } from './utils';
 
 import './CreateCommunity.scss';
 
 const CreateCommunity = () => {
-  const [createCommunityStep, setCreateCommunityStep] =
-    useState<CreateCommunityStep>(CreateCommunityStep.CommunityTypeSelection);
-  const [selectedCommunity, setSelectedCommunity] = useState<SelectedCommunity>(
-    { type: null, chainBase: null },
-  );
-  const [selectedAddress, setSelectedAddress] = useState<AddressInfo>(null);
-  const [selectedChainId, setSelectedChainId] = useState(null);
-  const [createdCommunityId, setCreatedCommunityId] = useState('');
-  const [createdCommunityName, setCreatedCommunityName] = useState('');
+  const {
+    createCommunityStep,
+    selectedCommunity,
+    setSelectedCommunity,
+    selectedAddress,
+    setSelectedAddress,
+    setSelectedChainId,
+    createdCommunityId,
+    createdCommunityName,
+    handleCompleteBasicInformationStep,
+    onChangeStep,
+    showCommunityStakeStep,
+  } = useCreateCommunity();
 
   useBrowserAnalyticsTrack({
     payload: { event: MixpanelCommunityCreationEvent.CREATE_COMMUNITY_VISITED },
   });
 
   const isSuccessStep = createCommunityStep === CreateCommunityStep.Success;
-
-  const isValidStepToShowCommunityStakeFormStep = [
-    CreateCommunityStep.BasicInformation,
-    CreateCommunityStep.CommunityStake,
-  ].includes(createCommunityStep);
-  const isEthereumMainnetSelected = selectedChainId === ETHEREUM_MAINNET_ID;
-  const showCommunityStakeStep =
-    isValidStepToShowCommunityStakeFormStep &&
-    selectedCommunity.type === 'ethereum' &&
-    isEthereumMainnetSelected;
-
-  const handleCompleteBasicInformationStep = (
-    communityId: string,
-    communityName: string,
-  ) => {
-    onChangeStep(true);
-    setCreatedCommunityId(communityId);
-    setCreatedCommunityName(communityName);
-  };
-
-  const onChangeStep = (forward: boolean) => {
-    handleChangeStep(
-      forward,
-      createCommunityStep,
-      setCreateCommunityStep,
-      showCommunityStakeStep,
-    );
-  };
 
   const getCurrentStep = () => {
     switch (createCommunityStep) {
